Delete brand-visibility-worker-template && update ai-brand-visibility-template
