+++ ai-brand-visibility-template/.gitignore
@@ -5,3 +5,4 @@ dist/
 build/
 .backup-vanilla/
 public/_old-*
+worker-configuration.d.ts
