+++ ai-brand-visibility-template/src/config.ts
@@ -17,13 +17,15 @@ export type ModelConfig = {
 		| "openai"
 		| "anthropic"
 		| "google"
-		| "workers-ai"
 		| "meta"
-		| "mistral";
+		| "mistral"
+		| "workers-ai";
 	/** Max tokens to request */
 	maxTokens?: number;
 	/** Whether this is a Gemini model (different request/response format) */
 	isGemini?: boolean;
+	/** Whether this is an Anthropic model (no system role in messages) */
+	isAnthropic?: boolean;
 };
 
 /**
@@ -43,6 +45,7 @@ export const MODELS: ModelConfig[] = [
 		name: "Claude Sonnet 4",
 		provider: "anthropic",
 		maxTokens: 512,
+		isAnthropic: true,
 	},
 	{
 		id: "google/gemini-3-flash",
