--- brand-visibility-worker-template/.gitignore
@@ -1,7 +1,0 @@
-node_modules/
-.wrangler/
-.react-router/
-dist/
-build/
-.backup-vanilla/
-public/_old-*
