--- brand-visibility-worker-template/package.json
@@ -1,45 +1,0 @@
-{
-	"name": "brand-visibility-worker-template",
-	"description": "Test whether AI models mention your brand when answering relevant queries. Runs prompts through GPT-5.4, Claude Sonnet 4, Gemini 3 Flash, Llama 4, and Mistral via Cloudflare AI Gateway.",
-	"private": true,
-	"cloudflare": {
-		"label": "Brand Visibility Worker Template",
-		"products": [
-			"Workers"
-		],
-		"categories": [
-			"ai"
-		],
-		"docs_url": "https://developers.cloudflare.com/ai-gateway/"
-	},
-	"type": "module",
-	"scripts": {
-		"dev": "react-router dev",
-		"build": "react-router build",
-		"deploy": "npm run build && wrangler deploy",
-		"preview": "npm run build && vite preview",
-		"cf-typegen": "wrangler types",
-		"typecheck": "npm run cf-typegen && react-router typegen && tsc -b"
-	},
-	"dependencies": {
-		"@phosphor-icons/react": "^2.1.10",
-		"hono": "4.11.1",
-		"isbot": "5.1.32",
-		"react": "19.2.1",
-		"react-dom": "19.2.1",
-		"react-router": "7.9.6"
-	},
-	"devDependencies": {
-		"@cloudflare/vite-plugin": "1.15.3",
-		"@cloudflare/workers-types": "^4.20250521.0",
-		"@react-router/dev": "7.9.6",
-		"@tailwindcss/vite": "4.1.17",
-		"@types/react": "19.2.7",
-		"@types/react-dom": "19.2.3",
-		"tailwindcss": "4.1.17",
-		"typescript": "5.9.3",
-		"vite": "^6.0.0",
-		"vite-tsconfig-paths": "5.1.4",
-		"wrangler": "^4.17.0"
-	}
-}
