--- brand-visibility-worker-template/tsconfig.json
@@ -1,15 +1,0 @@
-{
-	"files": [],
-	"references": [
-		{ "path": "./tsconfig.node.json" },
-		{ "path": "./tsconfig.cloudflare.json" }
-	],
-	"compilerOptions": {
-		"checkJs": true,
-		"verbatimModuleSyntax": true,
-		"skipLibCheck": true,
-		"strict": true,
-		"noEmit": true,
-		"types": ["@cloudflare/workers-types/2023-07-01"]
-	}
-}
