--- brand-visibility-worker-template/tsconfig.node.json
@@ -1,13 +1,0 @@
-{
-	"extends": "./tsconfig.json",
-	"include": ["vite.config.ts"],
-	"compilerOptions": {
-		"composite": true,
-		"strict": true,
-		"types": ["node"],
-		"lib": ["ES2022"],
-		"target": "ES2022",
-		"module": "ES2022",
-		"moduleResolution": "bundler"
-	}
-}
